Clean up persondetailinfo
